One major update
While this is NOT the way to commit changes, I didn't had the actual time to split everything into multiple commits. I wrote the code on the road during the weekend, whenever I'd get some spare time.

The purpose of this was to add a store and use tailwind. As my intention was to not plug in a store but to give it some sort of actual purpose, I had to make the application more complicated.

This massive update includes:
- renaming from ToDo to Tasks, because it was annoying to spell
- added Pinia Store as a single source of truth
- moved endpoints in the store
- added vue router for routing between the pages of the SPA
- redesigned the entire interface
- switched from written CSS to building the style with Tailwind, based on its helper classes